resource/aws_elasticsearch_domain: aws conversion methods; schema validation
