+++ aws/structure.go
@@ -1213,48 +1213,6 @@ func expandTxtEntry(s string) string {
 	return s
 }
 
-func expandAdvancedSecurityOptions(m []interface{}) *elasticsearch.AdvancedSecurityOptionsInput {
-	config := elasticsearch.AdvancedSecurityOptionsInput{}
-	group := m[0].(map[string]interface{})
-
-	if v, ok := group["enabled"].(bool); ok {
-		config.Enabled = aws.Bool(v)
-	}
-
-	if v, ok := group["internal_user_database_enabled"].(bool); ok {
-		config.InternalUserDatabaseEnabled = aws.Bool(v)
-	}
-
-	if v, ok := group["master_user_options"].([]interface{}); ok {
-		if len(v) > 0 {
-			muo := elasticsearch.MasterUserOptions{}
-			masterUserOptions := v[0].(map[string]interface{})
-
-			if v, ok := masterUserOptions["master_user_arn"].(string); ok {
-				if v != "" {
-					muo.MasterUserARN = aws.String(v)
-				}
-			}
-
-			if v, ok := masterUserOptions["master_user_name"].(string); ok {
-				if v != "" {
-					muo.MasterUserName = aws.String(v)
-				}
-			}
-
-			if v, ok := masterUserOptions["master_user_password"].(string); ok {
-				if v != "" {
-					muo.MasterUserPassword = aws.String(v)
-				}
-			}
-
-			config.SetMasterUserOptions(&muo)
-		}
-	}
-
-	return &config
-}
-
 func expandESClusterConfig(m map[string]interface{}) *elasticsearch.ElasticsearchClusterConfig {
 	config := elasticsearch.ElasticsearchClusterConfig{}
 
@@ -1429,19 +1387,6 @@ func flattenESSnapshotOptions(snapshotOptions *elasticsearch.SnapshotOptions) []
 
 	m := map[string]interface{}{
 		"automated_snapshot_start_hour": int(aws.Int64Value(snapshotOptions.AutomatedSnapshotStartHour)),
-	}
-
-	return []map[string]interface{}{m}
-}
-
-func flattenAdvancedSecurityOptions(advancedSecurityOptions *elasticsearch.AdvancedSecurityOptions) []map[string]interface{} {
-	if advancedSecurityOptions == nil {
-		return []map[string]interface{}{}
-	}
-
-	m := map[string]interface{}{
-		"enabled":                        *advancedSecurityOptions.Enabled,
-		"internal_user_database_enabled": *advancedSecurityOptions.InternalUserDatabaseEnabled,
 	}
 
 	return []map[string]interface{}{m}
